requirements.txt
numpy==1.24.3
pandas==1.5.3
Pillow==9.4.0

plotly==5.18.0

Requests==2.31.0
streamlit==1.31.0
streamlit_lottie==0.0.5